requirements.txt
# Added for Cyber AIops prototype
scikit-learn>=1.2
pandas>=1.5
joblib
scipy
PyPDF2>=3.0
pypdf>=3.0
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
# Embedding based auto-tagging
sentence-transformers>=2.2.2
faiss-cpu
numpy
